Split behemoth ResourceRequestsSpec
